+++ crates/sotf-plugins/crates/sotf-plugin-channel-mute-solo/src/lib.rs
@@ -3,10 +3,6 @@ pub mod params;
 
 #[path = "lib/channel_mute_solo_plugin.rs"]
 mod channel_mute_solo_plugin;
-#[path = "lib/default.rs"]
-mod default;
-#[path = "lib/misc.rs"]
-mod misc;
 #[cfg(test)]
 #[path = "lib/tests.rs"]
 mod tests;